pii-manager/requirements.txt
python-stdnum >=1.17,<2.0
